+++ GenericPM-Spores/include/organS.cpp
@@ -21,7 +21,7 @@ void OrganS::integrationS() {
     CloudOS *cloudo;
     LesionCohortS *lc;
 
-    float cloudDensity = 0;
+    double cloudDensity = 0;
     double cloudOValue = 0, cloudPValue = 0, cloudFvalue = 0;
     dailyDiseaseArea = diseaseArea;
     dailyVisibleDiseaseArea = visibleDiseaseArea;
@@ -124,8 +124,8 @@ void OrganS::cloudIntegrationS() {
         (&cloudsO[i])->integrationS();
     }
 }
-float OrganS::cloudAmountS() {
-    float cloudOValue=0;
+double OrganS::cloudAmountS() {
+    double cloudOValue=0;
     for (unsigned int i = 0; i < cloudsO.size(); i++) {
         cloudOValue += (&cloudsO[i])->getValueS();
     }
@@ -159,7 +159,7 @@ void OrganS::outputS() {
 
 void OrganS::rateS() {
     // Calculate the ratio due senescence based on previews day
-    float actualDisease=0, ratioSenescence = this->senescenceArea / this->totalArea;
+    double actualDisease=0, ratioSenescence = this->senescenceArea / this->totalArea;
     // Update the senescence area for the current day
     this->senescenceArea = SimulatorS::getInstanceS()->getCropInterface()->getSenescenceOrganArea(organNumber);
     // Recalculate the ratio due senescence and take the difference from previews ratio
